refactor(runtime): split review-loop drift handling into focused modules
