+++ src/crewplane/runtime/execution/review_loop/drift/__init__.py
@@ -1,0 +1,8 @@
+"""Guard review-loop provider calls against artifact drift."""
+
+from .guard import create_drift_guard_session, run_provider_call_with_drift_guard
+
+__all__ = [
+    "create_drift_guard_session",
+    "run_provider_call_with_drift_guard",
+]
