+++ src/crewplane/runtime/execution/review_loop/drift/events.py
@@ -6,8 +6,8 @@ from crewplane.architecture.ports import ArtifactStorePort
 from crewplane.core.preflight.models import ProviderRecord
 from crewplane.core.workflow.keywords import ProviderRole
 
-from ..common import ExecutionTelemetry, RuntimeEventContext, emit_runtime_log
-from .types import DriftCheckResult
+from ...common import ExecutionTelemetry, RuntimeEventContext, emit_runtime_log
+from ..types import DriftCheckResult
 
 
 def path_relative_to_state_root(
